+++ quill/include/quill/handlers/JsonFileHandler.h
@@ -14,8 +14,8 @@ class JsonFileHandler : public FileHandler
 {
 public:
   JsonFileHandler(fs::path const& filename, std::string const& mode, FilenameAppend append_to_filename,
-                  FileEventNotifier file_event_notifier = FileEventNotifier{})
-    : FileHandler(filename, mode, append_to_filename, std::move(file_event_notifier))
+                  FileEventNotifier file_event_notifier, bool do_fsync)
+    : FileHandler(filename, mode, append_to_filename, std::move(file_event_notifier), do_fsync)
   {
     // JsonFileHandler requires an empty pattern
     set_pattern("");