+++ quill/test/HandlerCollectionTest.cpp
@@ -20,7 +20,7 @@ TEST_CASE("stdout_stderr_handlers")
 
   // Attempt to create a file handler with stdout as name and check that it is the same as the default
   StreamHandler* filehandler_1 =
-    hc.create_handler<FileHandler>("stdout", "a", FilenameAppend::None, FileEventNotifier{});
+    hc.create_handler<FileHandler>("stdout", "a", FilenameAppend::None, FileEventNotifier{}, false);
   REQUIRE_EQ(filehandler_1, stdout_handler);
 
   // Get the default stderr stream handler
@@ -29,7 +29,7 @@ TEST_CASE("stdout_stderr_handlers")
 
   // Attempt to create a file handler with stderr as name and check that it is the same as the default
   StreamHandler* filehandler_2 =
-    hc.create_handler<FileHandler>("stderr", "a", FilenameAppend::None, FileEventNotifier{});
+    hc.create_handler<FileHandler>("stderr", "a", FilenameAppend::None, FileEventNotifier{}, false);
   REQUIRE_EQ(filehandler_2, stderr_handler);
 }
 
@@ -40,11 +40,11 @@ TEST_CASE("create_get")
 
   // Create a file handler
   StreamHandler* filehandler = hc.create_handler<FileHandler>(
-    "create_get_file_handler", "w", FilenameAppend::None, FileEventNotifier{});
+    "create_get_file_handler", "w", FilenameAppend::None, FileEventNotifier{}, false);
 
   // Request the same file handler
   StreamHandler* filehandler_2 = hc.create_handler<FileHandler>(
-    "create_get_file_handler", "a", FilenameAppend::None, FileEventNotifier{});
+    "create_get_file_handler", "a", FilenameAppend::None, FileEventNotifier{}, false);
 
   // Compare the pointers
   REQUIRE_EQ(filehandler, filehandler_2);
@@ -57,11 +57,11 @@ TEST_CASE("subscribe_get_active_same_handler")
   HandlerCollection hc;
   // Create a file handler
   StreamHandler* filehandler = hc.create_handler<FileHandler>(
-    "create_get_file_handler", "w", FilenameAppend::None, FileEventNotifier{});
+    "create_get_file_handler", "w", FilenameAppend::None, FileEventNotifier{}, false);
 
   // Request the same file handler
   StreamHandler* filehandler_2 = hc.create_handler<FileHandler>(
-    "create_get_file_handler", "a", FilenameAppend::None, FileEventNotifier{});
+    "create_get_file_handler", "a", FilenameAppend::None, FileEventNotifier{}, false);
 
   // Compare the pointers
   REQUIRE_EQ(filehandler, filehandler_2);
@@ -94,11 +94,11 @@ TEST_CASE("subscribe_get_active_different_handlers")
 
   // Create a file handler
   StreamHandler* filehandler = hc.create_handler<FileHandler>(
-    "create_get_file_handler_1", "w", FilenameAppend::None, FileEventNotifier{});
+    "create_get_file_handler_1", "w", FilenameAppend::None, FileEventNotifier{}, false);
 
   // Request the same file handler
   StreamHandler* filehandler_2 = hc.create_handler<FileHandler>(
-    "create_get_file_handler_2", "a", FilenameAppend::None, FileEventNotifier{});
+    "create_get_file_handler_2", "a", FilenameAppend::None, FileEventNotifier{}, false);
 
   // Compare the pointers
   REQUIRE_NE(filehandler, filehandler_2);