+++ quill/test/LogTest.cpp
@@ -40,7 +40,7 @@ TEST_CASE("default_logger_with_filehandler_1")
   // Set a file handler as the custom logger handler and log to it
   quill::Config cfg;
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false));
   lm.configure(cfg);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -83,7 +83,7 @@ TEST_CASE("default_logger_with_filehandler_2")
 
   quill::Config cfg;
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false));
   lm.configure(cfg);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -127,7 +127,7 @@ TEST_CASE("default_logger_ints_and_large_string")
 
   quill::Config cfg;
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false));
   lm.configure(cfg);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -173,7 +173,7 @@ TEST_CASE("default_logger_ints_and_very_large_string")
 
   quill::Config cfg;
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false));
   lm.configure(cfg);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -228,7 +228,7 @@ TEST_CASE("default_logger_ints_and_wide_string")
 
   quill::Config cfg;
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false));
   lm.configure(cfg);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -266,7 +266,7 @@ TEST_CASE("default_logger_ints_and_very_large_wide_string")
 
   quill::Config cfg;
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false));
   lm.configure(cfg);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -319,7 +319,7 @@ void custom_default_logger_same_handler(int test_case = 0)
 
   // Set a file handler the custom logger handler and log to it
   Handler* file_handler = lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{});
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false);
   file_handler->set_pattern("%(ascii_time) %(logger_name) - %(message) [%(level_id)]");
 
   quill::Config cfg;
@@ -333,7 +333,7 @@ void custom_default_logger_same_handler(int test_case = 0)
   {
     // Add a second logger using the same file handler
     Handler* file_handler_2 = lm.handler_collection().create_handler<FileHandler>(
-      filename.string(), "a", FilenameAppend::None, FileEventNotifier{});
+      filename.string(), "a", FilenameAppend::None, FileEventNotifier{}, false);
     QUILL_MAYBE_UNUSED Logger* logger_2 =
       lm.create_logger("custom_logger", file_handler_2, std::nullopt, std::nullopt);
   }
@@ -416,12 +416,12 @@ void test_custom_default_logger_multiple_handlers(int test_case)
 
   // First handler
   Handler* file_handler_1 = lm.handler_collection().create_handler<FileHandler>(
-    filename_1.string(), "w", FilenameAppend::None, FileEventNotifier{});
+    filename_1.string(), "w", FilenameAppend::None, FileEventNotifier{}, false);
   file_handler_1->set_pattern("%(ascii_time) %(logger_name) - %(message) [%(level_id)]");
 
   // Second handler with different pattern
   Handler* file_handler_2 = lm.handler_collection().create_handler<FileHandler>(
-    filename_2.string(), "w", FilenameAppend::None, FileEventNotifier{});
+    filename_2.string(), "w", FilenameAppend::None, FileEventNotifier{}, false);
   file_handler_2->set_pattern("%(ascii_time) %(logger_name) - %(message)", "%D %H:%M:%S.%Qms");
 
   quill::Config cfg;
@@ -436,9 +436,9 @@ void test_custom_default_logger_multiple_handlers(int test_case)
   {
     // Add a second logger using the same file handler
     Handler* file_handler_a = lm.handler_collection().create_handler<FileHandler>(
-      filename_1.string(), "", FilenameAppend::None, FileEventNotifier{});
+      filename_1.string(), "", FilenameAppend::None, FileEventNotifier{}, false);
     Handler* file_handler_b = lm.handler_collection().create_handler<FileHandler>(
-      filename_2.string(), "", FilenameAppend::None, FileEventNotifier{});
+      filename_2.string(), "", FilenameAppend::None, FileEventNotifier{}, false);
     QUILL_MAYBE_UNUSED Logger* logger_2 =
       lm.create_logger("custom_logger", {file_handler_a, file_handler_b}, std::nullopt, std::nullopt);
   }
@@ -550,7 +550,7 @@ TEST_CASE("many_loggers_multiple_threads")
 
   quill::Config cfg;
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}, false));
   lm.configure(cfg);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -638,7 +638,7 @@ TEST_CASE("backend_error_handler")
 
       // Set a file handler as the custom logger handler and log to it
       cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-        filename.string(), "a", FilenameAppend::None, FileEventNotifier{}));
+        filename.string(), "a", FilenameAppend::None, FileEventNotifier{}, false));
 
       // Start backend worker
       lm.configure(cfg);
@@ -686,7 +686,7 @@ TEST_CASE("backend_error_handler_log_from_backend_thread")
 
       // Set a file handler as the custom logger handler and log to it
       cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-        filename.string(), "a", FilenameAppend::None, FileEventNotifier{}));
+        filename.string(), "a", FilenameAppend::None, FileEventNotifier{}, false));
 
       // Set a custom error handler to handler exceptions
       cfg.backend_thread_error_handler = [&lm](std::string const& s)
@@ -737,7 +737,7 @@ TEST_CASE("backend_error_handler_error_throw_while_in_backend_process")
 
   // Set a file handler as the custom logger handler and log to it
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}, false));
 
   // Set a custom error handler to handler exceptions
   cfg.backend_thread_error_handler = [&error_handler_invoked](std::string const& s)
@@ -789,7 +789,7 @@ TEST_CASE("log_backtrace_and_flush_on_error_1")
 
   // Set a file handler as the custom logger handler and log to it
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}, false));
 
   lm.configure(cfg);
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -846,7 +846,7 @@ TEST_CASE("log_backtrace_and_flush_on_error_2")
 
   // Set a file handler as the custom logger handler and log to it
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}, false));
 
   lm.configure(cfg);
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -909,7 +909,7 @@ TEST_CASE("log_backtrace_terminate_thread_then_and_flush_on_error")
 
   // Set a file handler as the custom logger handler and log to it
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}, false));
 
   lm.configure(cfg);
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -975,7 +975,7 @@ TEST_CASE("log_backtrace_manual_flush")
 
   // Set a file handler as the custom logger handler and log to it
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "a", FilenameAppend::None, FileEventNotifier{}, false));
 
   lm.configure(cfg);
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -1093,14 +1093,14 @@ TEST_CASE("logger_with_two_files_filters")
 
   // Set file 1
   quill::Handler* file_handler1 = lm.handler_collection().create_handler<FileHandler>(
-    filename1.string(), "w", FilenameAppend::None, FileEventNotifier{});
+    filename1.string(), "w", FilenameAppend::None, FileEventNotifier{}, false);
 
   // Create and add the filter to our handler
   file_handler1->add_filter(std::make_unique<FileFilter1>());
 
   // Set file 2
   quill::Handler* file_handler2 = lm.handler_collection().create_handler<FileHandler>(
-    filename2.string(), "w", FilenameAppend::None, FileEventNotifier{});
+    filename2.string(), "w", FilenameAppend::None, FileEventNotifier{}, false);
 
   // Create and add the filter to our handler
   file_handler2->add_filter(std::make_unique<FileFilter2>());
@@ -1150,11 +1150,11 @@ TEST_CASE("logger_with_two_files_set_log_level_on_handler")
 
   // Set file 1
   quill::Handler* file_handler1 = lm.handler_collection().create_handler<FileHandler>(
-    filename1.string(), "w", FilenameAppend::None, FileEventNotifier{});
+    filename1.string(), "w", FilenameAppend::None, FileEventNotifier{}, false);
 
   // Set file 2
   quill::Handler* file_handler2 = lm.handler_collection().create_handler<FileHandler>(
-    filename2.string(), "w", FilenameAppend::None, FileEventNotifier{});
+    filename2.string(), "w", FilenameAppend::None, FileEventNotifier{}, false);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
 
@@ -1250,7 +1250,7 @@ TEST_CASE("default_logger_with_custom_timestamp")
 
   // also change the pattern to log the date
   Handler* handler = lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{});
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false);
   handler->set_pattern("%(ascii_time) %(level_name) %(logger_name:<16) - %(message)", // format
                        "%Y-%m-%d %H:%M:%S.%Qms",  // timestamp format
                        quill::Timezone::GmtTime); // timestamp's timezone
@@ -1318,40 +1318,44 @@ TEST_CASE("log_configure_default_logger_single_handler")
   // configure a custom handler to the default logger and check it works
   fs::path const filename = "log_configure_default_logger_single_handler.log";
 
-  LogManager lm;
+  {
+    // LogManager needs to go out of score for the files to be closed and FileEventNotifier get
+    // called on file close
+    LogManager lm;
 
-  // first obtain the pointer before configure
-  Logger* default_logger = lm.logger_collection().get_logger();
+    // first obtain the pointer before configure
+    Logger* default_logger = lm.logger_collection().get_logger();
 
-  // Config using the custom ts class and the stdout handler
-  // this should not invalidate the default logger
-  quill::Config cfg;
-  Handler* handler = lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{});
-  cfg.default_handlers.emplace_back(handler);
-  cfg.default_logger_name = "root_test";
-  lm.configure(cfg);
+    // Config using the custom ts class and the stdout handler
+    // this should not invalidate the default logger
+    quill::Config cfg;
+    Handler* handler = lm.handler_collection().create_handler<FileHandler>(
+      filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, true);
+    cfg.default_handlers.emplace_back(handler);
+    cfg.default_logger_name = "root_test";
+    lm.configure(cfg);
 
-  // Start the backend logging thread
-  lm.start_backend_worker(false, std::initializer_list<int32_t>{});
+    // Start the backend logging thread
+    lm.start_backend_worker(false, std::initializer_list<int32_t>{});
 
-  std::thread frontend(
-    [&lm, &default_logger, &cfg]()
-    {
-      // check that we did update the default logger
-      auto all_loggers = lm.logger_collection().get_all_loggers();
-      REQUIRE_EQ(all_loggers.size(), 1u);
-      REQUIRE_NE(all_loggers.find(cfg.default_logger_name), all_loggers.end());
+    std::thread frontend(
+      [&lm, &default_logger, &cfg]()
+      {
+        // check that we did update the default logger
+        auto all_loggers = lm.logger_collection().get_all_loggers();
+        REQUIRE_EQ(all_loggers.size(), 1u);
+        REQUIRE_NE(all_loggers.find(cfg.default_logger_name), all_loggers.end());
 
-      // Log using the default logger pointer we obtained before the configure step
-      // We expect the pointer to be valid after quill::configure(...)
-      LOG_INFO(default_logger, "Hello from log_configure_default_logger test");
-      LOG_INFO(default_logger, "The root logger pointer is valid");
-    });
+        // Log using the default logger pointer we obtained before the configure step
+        // We expect the pointer to be valid after quill::configure(...)
+        LOG_INFO(default_logger, "Hello from log_configure_default_logger test");
+        LOG_INFO(default_logger, "The root logger pointer is valid");
+      });
 
-  frontend.join();
+    frontend.join();
 
-  lm.stop_backend_worker();
+    lm.stop_backend_worker();
+  }
 
   // Read file and check
   std::vector<std::string> const file_contents = quill::testing::file_contents(filename);
@@ -1396,9 +1400,9 @@ TEST_CASE("log_configure_default_logger_multi_handler")
     // this should not invalidate the default logger
     quill::Config cfg;
     quill::Handler* file_handler_1 = lm.handler_collection().create_handler<FileHandler>(
-      filename_1.string(), "w", FilenameAppend::None, FileEventNotifier{});
+      filename_1.string(), "w", FilenameAppend::None, FileEventNotifier{}, false);
     quill::Handler* file_handler_2 = lm.handler_collection().create_handler<FileHandler>(
-      filename_2.string(), "w", FilenameAppend::None, file_event_notifier);
+      filename_2.string(), "w", FilenameAppend::None, file_event_notifier, false);
     cfg.default_handlers.emplace_back(file_handler_1);
     cfg.default_handlers.emplace_back(file_handler_2);
     cfg.default_logger_name = "root_test";
@@ -1505,7 +1509,7 @@ TEST_CASE("default_logger_with_random_strings")
   // Set a file handler as the custom logger handler and log to it
   quill::Config cfg;
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false));
   lm.configure(cfg);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -1591,7 +1595,7 @@ TEST_CASE("default_logger_with_small_random_strings")
   // Set a file handler as the custom logger handler and log to it
   quill::Config cfg;
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false));
   lm.configure(cfg);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});
@@ -1677,7 +1681,7 @@ TEST_CASE("default_logger_with_very_large_random_strings")
   // Set a file handler as the custom logger handler and log to it
   quill::Config cfg;
   cfg.default_handlers.emplace_back(lm.handler_collection().create_handler<FileHandler>(
-    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}));
+    filename.string(), "w", FilenameAppend::None, FileEventNotifier{}, false));
   lm.configure(cfg);
 
   lm.start_backend_worker(false, std::initializer_list<int32_t>{});